Rename file to match PR's file name
